fixed: schema 위치 수정
