+++ protocol/review.go
@@ -1,25 +1,5 @@
 package protocol
 
-import (
-	"time"
-
-	"go.mongodb.org/mongo-driver/bson/primitive"
-)
-
-// review post info schema
-type ReviewPost struct {
-	Id        primitive.ObjectID `json:"id" bson:"_id"`
-	Title     string             `json:"title" bson:"title"`
-	UserId    string             `json:"userId" bson:"user_id"`
-	Content   string             `json:"content" bson:"content"`
-	Likes     int                `json:"likes" bson:"likes"`
-	LikeUsers []string           `json:"likeUsers" bson:"like_users"`
-	CreateAt  time.Time          `json:"createAt" bson:"create_at"`
-	UpdateAt  time.Time          `json:"updateAt" bson:"update_at"`
-	Stat      int                `json:"stat" bson:"stat"`
-	Grade     float64            `json:"grade" bson:"grade"`
-}
-
 type ReviewListResp struct {
 	*RespHeader
 	ReviewList []ReviewPost `json:"reviewList"`
