+++ protocol/types.go
@@ -1,6 +1,7 @@
 package protocol
 
 import (
+	"go.mongodb.org/mongo-driver/bson/primitive"
 	"strings"
 	"time"
 )
@@ -32,4 +33,18 @@ type PostInfo struct {
 	CommentId []int     `json:"commentId" bson:"comment_id"`
 	FileId    int       `json:"fileId" bson:"file_id"`
 	Stat      int       `json:"stat" bson:"stat"`
+}
+
+// review post info schema
+type ReviewPost struct {
+	Id        primitive.ObjectID `json:"id" bson:"_id"`
+	Title     string             `json:"title" bson:"title"`
+	UserId    string             `json:"userId" bson:"user_id"`
+	Content   string             `json:"content" bson:"content"`
+	Likes     int                `json:"likes" bson:"likes"`
+	LikeUsers []string           `json:"likeUsers" bson:"like_users"`
+	CreateAt  time.Time          `json:"createAt" bson:"create_at"`
+	UpdateAt  time.Time          `json:"updateAt" bson:"update_at"`
+	Stat      int                `json:"stat" bson:"stat"`
+	Grade     float64            `json:"grade" bson:"grade"`
 }
